requirements.txt
absl-py==0.11.0
aiohttp==3.5.4
alabaster==0.7.12
alembic==1.3.1
amqp==2.5.2
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
aniso8601==8.0.0
apache-airflow==1.10.12
apispec==1.3.3
appnope==0.1.0
appscript==1.0.1
argcomplete==1.12.0
arrow==0.15.4
asn1crypto==0.24.0
astor==0.8.1
astroid==2.2.5
astropy==3.2.1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.7.1
bert-for-tf2==0.14.6
billiard==3.6.0.0
binaryornot==0.4.4
bitarray==0.9.3
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.2.0
boto==2.49.0
boto3==1.10.27
botocore==1.13.27
Bottleneck==1.2.1
bs4==0.0.1
cached-property==1.5.1
cachetools==4.1.1
cairocffi==1.0.2
CairoSVG==2.3.0
cattrs==1.0.0
celery==4.3.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
clickhouse-driver==0.1.0
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.1
colorlog==4.0.2
conda==4.8.3
conda-build==3.18.8
conda-pack==0.4.0
conda-package-handling==1.3.11
conda-verify==3.4.2
configparser==3.5.3
contextlib2==0.5.5
cookiecutter==1.7.0
croniter==0.3.34
cryptography==2.7
cssselect2==0.2.1
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.1.0
databricks-cli==0.9.1
decorator==4.4.0
defusedxml==0.5.0
dgl==0.3.1
dill==0.2.9
distributed==2.1.0
dnspython==1.16.0
docker==4.1.0
docutils==0.14
dominate==2.4.0
dpath==1.4.2
email-validator==1.1.1
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.29.1
faiss-cpu==1.7.0
fastcache==1.1.0
fasttext==0.9.2
featuretools==0.12.0
filelock==3.0.12
flasgger==0.9.3
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.3.3
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
flower==0.9.5
fsspec==0.8.5
funcsigs==1.0.2
future==0.17.1
gast==0.3.3
get==2019.4.13
gevent==1.4.0
gitdb2==2.0.6
GitPython==3.0.5
glob2==0.7
gmpy2==2.0.8
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gorilla==0.3.0
graphframes==0.6
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.34.0
gunicorn==20.0.2
h5py==2.10.0
hanlp==2.0.0a67
happybase==1.2.0
heapdict==1.0.0
html5lib==1.0.1
htpasswd==2.3
humanize==3.1.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==1.7.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jieba==0.42.1
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
joblib==0.13.2
json-merge-patch==0.2
json5==0.8.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
kafka-python==2.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
kombu==4.6.3
lazy-object-proxy==1.4.1
libarchive-c==2.8
lief==0.9.0
lightgbm==2.2.3
line-profiler==2.1.2
llvmlite==0.29.0
locket==0.2.0
lockfile==0.12.2
lxml==4.3.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
-e git+https://git.creditx.com/marvin/marvin_modeling.git@d2dd966a276d8b4513d9e8467434aa1a11d0b8ac#egg=marvin_modeling
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mlflow==1.4.0
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.5.2
multipledispatch==0.6.0
multiprocess==0.70.8
mysqlclient==1.4.6
natsort==7.0.1
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
nbstripout==0.3.6
networkx==2.3
nltk==3.4.4
nose==1.3.7
notebook==6.0.0
numba==0.44.1
numexpr==2.6.9
numpy==1.18.0
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
openpyxl==2.6.2
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==19.0
paddlepaddle==2.0.1
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.6.0
params-flow==0.8.2
parso==0.5.0
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
pathos==0.2.4
patsy==0.5.1
pendulum==1.4.4
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
post==2019.4.13
pox==0.2.6
poyo==0.5.0
ppft==1.6.6.1
prison==0.1.3
prometheus-client==0.8.0
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.6.0
public==2019.4.13
py==1.8.0
py-params==0.9.7
py4j==0.10.9
pyalink===1.0.1-flink-1.9.0-scala-2.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
Pyphen==0.9.5
pyrsistent==0.14.11
PySocks==1.7.0
pyspark==3.0.0
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-crfsuite==0.9.7
python-daemon==2.2.4
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.3
python-nvd3==0.15.0
python-rocksdb==0.7.0
python-slugify==4.0.1
python-socketio==4.3.1
python3-openid==3.2.0
pytorch-lightning==1.1.6
pytz==2019.1
pytzdata==2020.1
PyWavelets==1.0.3
PyXB==1.2.6
PyYAML==5.1.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.8.0
query-string==2019.4.13
querystring-parser==1.2.4
redis==3.3.11
regex==2020.7.14
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.6
ruamel-yaml==0.15.46
s3fs==0.4.0
s3transfer==0.2.1
sacremoses==0.0.43
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.91
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.12.0
sklearn-crfsuite==0.3.6
sklearn-pandas==1.6.0
sklearn2pmml==0.35.15
smart-open==1.9.0
smmap2==2.0.5
snowballstemmer==1.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.0
statsmodels==0.10.0
sympy==1.4
tables==3.5.2
tabulate==0.8.6
tblib==1.4.0
tenacity==4.12.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
thrift==0.13.0
thriftpy2==0.4.11
tinycss2==0.6.1
tokenizers==0.10.1
toolz==0.10.0
torch==1.7.1
tornado==6.0.3
tqdm==4.56.0
traitlets==4.3.2
transformers==4.3.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
vine==1.3.0
visitor==0.1.3
wcwidth==0.1.7
WeasyPrint==45
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
whichcraft==0.6.1
widgetsnbextension==3.5.0
wrapt==1.11.2
WTForms==2.3.3
wurlitzer==1.0.2
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.1.8
xlwings==0.15.8
xlwt==1.3.0
xmltodict==0.12.0
xpinyin==0.5.6
yarl==1.3.0
zict==1.0.0
zipp==0.5.1
zope.deprecation==4.4.0
